service_1/requirements.txt
flask 
requests
flask_sqlalchemy
pymysql


